refactor: move `ConversationScope` and related classes to `model.scope` package
