+++ common/src/main/kotlin/io/fletchly/comparator/adapter/persistence/CaffeineContextStore.kt
@@ -23,7 +23,7 @@ import io.fletchly.comparator.model.message.Conversation
 import io.fletchly.comparator.model.message.Message
 import io.fletchly.comparator.model.message.conversationOf
 import io.fletchly.comparator.model.options.ContextOptions
-import io.fletchly.comparator.model.user.ConversationScope
+import io.fletchly.comparator.model.scope.ConversationScope
 import io.fletchly.comparator.port.out.ContextPort
 import java.util.UUID
 import java.util.concurrent.TimeUnit
