+++ common/src/main/kotlin/io/fletchly/comparator/adapter/persistence/InMemoryContextStore.kt
@@ -21,7 +21,7 @@ package io.fletchly.comparator.adapter.persistence
 import io.fletchly.comparator.model.message.Conversation
 import io.fletchly.comparator.model.message.Message
 import io.fletchly.comparator.model.message.conversationOf
-import io.fletchly.comparator.model.user.ConversationScope
+import io.fletchly.comparator.model.scope.ConversationScope
 import io.fletchly.comparator.port.out.ContextPort
 import io.fletchly.comparator.model.options.ContextOptions
 import kotlinx.coroutines.sync.Mutex
