+++ common/src/test/kotlin/io/fletchly/comparator/adapter/persistence/CaffeineContextStoreTest.kt
@@ -20,7 +20,7 @@ package io.fletchly.comparator.adapter.persistence
 
 import io.fletchly.comparator.model.message.Message
 import io.fletchly.comparator.model.message.conversationOf
-import io.fletchly.comparator.model.user.ConversationScope
+import io.fletchly.comparator.model.scope.ConversationScope
 import io.fletchly.comparator.model.options.ContextOptions
 import io.mockk.every
 import io.mockk.mockk
